[lib] Move init_elf_data() to readelf.c, move data to struct rpminspect
Rearranging things for the creation of a new ELF-related inspection.
The init_elf_data() function generates a list of all fortifiable
functions.  Cache this now in struct rpminspect rather than having it
be part of the elf inspection code as a static structure.  Remove the
free_elf_data() function since that's now handled with
free_rpminspect().
