+++ include/readelf.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2020  Red Hat, Inc.
+ * Copyright (C) 2019-2021  Red Hat, Inc.
  * Red Hat Author(s):  David Shea <dshea@redhat.com>
  *                     David Cantrell <dcantrell@redhat.com>
  *
@@ -37,6 +37,8 @@
 #define EM_BPF 247 /* Linux BPF -- in-kernel virtual machine */
 #endif
 
+bool is_fortified(const char *);
+void init_elf_data(struct rpminspect *);
 Elf *get_elf(const char *, int *);
 Elf *get_elf_archive(const char *, int *);
 GElf_Half get_elf_type(Elf *);
